backend/requirements.txt
requests==2.31.0
pymongo==4.6.0
schedule==1.2.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
marshmallow==3.20.1
psutil==5.9.6
feedparser==6.0.10
textblob==0.17.1
pytz==2024.1
PyJWT==2.8.0
werkzeug==3.0.1
redis==5.0.1

